requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
cadquery==2.5.2
cadquery-ocp==7.7.2
casadi==3.7.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
datasets==3.6.0
dill==0.3.8
ezdxf==1.4.2
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
huggingface-hub==0.33.0
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.6
loguru==0.7.3
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.4.4
multimethod==1.12
multiprocess==0.70.16
networkx==3.5
nlopt==2.9.1
numpy==2.3.1
omegaconf==2.3.0
packaging==25.0
pandas==2.3.0
path==17.1.0
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
sentencepiece==0.2.0
sentry-sdk==2.31.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sympy==1.14.0
tokenizers==0.19.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.40.2
typing-inspection==0.4.1
typing_extensions==4.14.0
typish==1.9.3
tzdata==2025.2
urllib3==2.5.0
wandb==0.20.1
win32_setctime==1.2.0
xxhash==3.5.0
yarl==1.20.1
